pyproject.toml
[project]
name = "ascii-soslo"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pyqt5 (>=5.15.11,<6.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "poetry-core (>=2.0.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "numba (>=0.62.1,<0.63.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
